PoC/controller/requirements.txt
certifi==2020.6.20
cryptography==3.3.1
paho-mqtt==1.6.1
Pyfhel==3.2.1
rsa==4.9